Split EventTemplate into event_template.py; rename from_record -> infer_from; consolidate to_numeric (#235)
Move EventTemplate / NumericEventTemplate, the leaf-spec sum type, and the
template helpers out of record.py into a foundational event_template.py module.
The import graph is now one-way (record -> event_template): EventTemplate's
value-facing conveniences (pack / to_vector / from_vector / infer_from /
_value_treedef) construct or consume Record values via function-local imports.

- Rename EventTemplate.from_record -> EventTemplate.infer_from, documented as a
  best-effort, lossy fallback (cannot recover support / opaque meta /
  Distribution / Function specs); the authoritative schema is carried, not
  re-inferred.
- Consolidate the Record -> NumericRecord conversion to Record.to_numeric();
  drop the colliding NumericRecord.from_record classmethod.
- _PATH_SEP / _check_no_path_sep now live in event_template.py (single home).
- Repoint all importers (~30 source + test modules) and probpipe/__init__.py.

Pure structural + rename refactor; no value-model change yet. Full suite green.
