+++ tests/core/test_numeric_record.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from probpipe import NumericRecord, Record
-from probpipe.core.record import EventTemplate
+from probpipe.core.event_template import EventTemplate
 
 # ---------------------------------------------------------------------------
 # Construction
@@ -193,7 +193,7 @@ class TestToVectorFromVector:
         np.testing.assert_allclose(float(nr2["phi"]), 10.0)
 
     def test_roundtrip_nested_template(self):
-        from probpipe.core.record import NumericEventTemplate
+        from probpipe.core.event_template import NumericEventTemplate
 
         inner_tpl = NumericEventTemplate(x=(), y=(2,))
         outer_tpl = NumericEventTemplate(params=inner_tpl, z=(3,))
@@ -215,21 +215,21 @@ class TestToVectorFromVector:
 
 
 # ---------------------------------------------------------------------------
-# from_record conversion
+# Record -> NumericRecord conversion (to_numeric)
 # ---------------------------------------------------------------------------
 
 
-class TestFromRecord:
+class TestToNumeric:
     def test_simple(self):
         r = Record(a=1.0, b=jnp.zeros(3))
-        nr = NumericRecord.from_record(r)
+        nr = r.to_numeric()
         assert isinstance(nr, NumericRecord)
         assert nr.fields == ("a", "b")
 
     def test_nested(self):
         inner = Record(x=1.0, y=2.0)
         outer = Record(params=inner, z=3.0)
-        nr = NumericRecord.from_record(outer)
+        nr = outer.to_numeric()
         assert isinstance(nr, NumericRecord)
         assert isinstance(nr["params"], NumericRecord)
